RAG/requirements.txt
fastapi
uvicorn
pypdf
chromadb
python-dotenv
google-genai
